requirements.txt
Django==1.8.2
djangorestframework==3.1.3
drf-nested-fields==0.9.2
drf-hal-json==0.9.0
drf-enum-field==0.9.1
drf-nested-routing==0.9.0
django-filter==0.10.0
openpyxl==2.2.5
chardet==2.3.0
